# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/healthcheck": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "sample query",
                        "name": "test",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "server is ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "server is not ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.BaseErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/user/{id}": {
            "post": {
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "server is ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "server is not ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.BaseErrorResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BaseErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer",
                    "example": 4
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "model.BaseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "integer"
                },
                "result": {
                    "type": "object"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "account_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "integer"
                },
                "ext_id": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "first_name_th": {
                    "type": "string"
                },
                "freeze": {
                    "type": "integer"
                },
                "freeze_crypto_withdraw": {
                    "type": "integer"
                },
                "google_authenticator": {
                    "type": "string"
                },
                "google_authenticator_trade_verified": {
                    "type": "integer"
                },
                "google_authenticator_verified": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_ip": {
                    "type": "string"
                },
                "last_login": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "last_name_th": {
                    "type": "string"
                },
                "last_pw_status": {
                    "type": "integer"
                },
                "last_pw_updated": {
                    "type": "string"
                },
                "mobile_pin": {
                    "type": "string"
                },
                "otp_enabled": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "rank_vip": {
                    "type": "integer"
                },
                "referral_id": {
                    "type": "integer"
                },
                "risk_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tel": {
                    "type": "string"
                },
                "trading_credit": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}